Split DictionaryAdapter into an external class so it can be used by the vocab sheet
